updates for results.json
