+++ sdk/evaluation/azure-ai-evaluation/azure/ai/evaluation/red_team/_red_team_result.py
@@ -340,9 +340,9 @@ class ScanResult(TypedDict):
     :param attack_details: List of AttackDetails objects representing the conversations in the evaluation
     :type attack_details: List[AttackDetails]
     :param AOAI_Compatible_Row_Results: List of evaluation results for each risk category
-    :type AOAI_Compatible_Row_Results: List[RedTeamRunOutputItemResult]
+    :type AOAI_Compatible_Row_Results: Optional[List[RedTeamRunOutputItemResult]]
     :param AOAI_Compatible_Summary: The evaluation run metadata in eval.run format
-    :type AOAI_Compatible_Summary: RedTeamRun
+    :type AOAI_Compatible_Summary: Optional[RedTeamRun]
     :param studio_url: Optional URL for the studio
     :type studio_url: Optional[str]
     """
@@ -350,8 +350,8 @@ class ScanResult(TypedDict):
     scorecard: RedTeamingScorecard
     parameters: RedTeamingParameters
     attack_details: List[AttackDetails]
-    AOAI_Compatible_Row_Results: List[RedTeamRunOutputItemResult]
-    AOAI_Compatible_Summary: "RedTeamRun"
+    AOAI_Compatible_Row_Results: Optional[List[RedTeamRunOutputItemResult]]
+    AOAI_Compatible_Summary: Optional["RedTeamRun"]
     studio_url: Optional[str]
